kes  Make bscan ignore ACL streams. kes  Update projects files with new Feature Requests. kes  Generate current English and French manuals and upload      them to the site. kes  Update the home page to include the new British/French flags to      switch between translations. Work done by Alexandre Baron. kes  Make disk-changer script print an error message if the changer      directory is not defined. kes  Create a generic DB routine for getting an int (32/64 bit) from      the DB. kes  Add more migration code. kes  Add a few more files to .cvsignore here and there.
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@3201 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/filed/job.c
@@ -1328,14 +1328,14 @@ static int backup_cmd(JCR *jcr)
             } else {
                /* tell user if snapshot creation of a specific drive failed */
                int i;
-               for (i=0; i < strlen(szWinDriveLetters); i++) {
+               for (i=0; i < (int)strlen(szWinDriveLetters); i++) {
                   if (islower(szWinDriveLetters[i])) {
                      Jmsg(jcr, M_WARNING, 0, _("Generate VSS snapshot of drive \"%c:\\\" failed. VSS support is disabled on this drive.\n"), szWinDriveLetters[i]);
                      jcr->Errors++;
                   }
                }
                /* inform user about writer states */
-               for (i=0; i<g_pVSSClient->GetWriterCount(); i++)                
+               for (i=0; i < (int)g_pVSSClient->GetWriterCount(); i++)                
                   if (g_pVSSClient->GetWriterState(i) < 1) {
                      Jmsg(jcr, M_WARNING, 0, _("VSS Writer (PrepareForBackup): %s\n"), g_pVSSClient->GetWriterInfo(i));                    
                      jcr->Errors++;
